Extract PaginatedMultiSelect to its own module
